move FeedingPhotoItem to the model package
