+++ feature/home/src/main/java/com/epmedu/animeal/home/presentation/viewmodel/handlers/gallery/DefaultFeedingPhotoGalleryHandler.kt
@@ -3,7 +3,7 @@ package com.epmedu.animeal.home.presentation.viewmodel.handlers.gallery
 import com.epmedu.animeal.camera.domain.usecase.DeletePhotoUseCase
 import com.epmedu.animeal.common.presentation.viewmodel.delegate.ActionDelegate
 import com.epmedu.animeal.common.presentation.viewmodel.delegate.StateDelegate
-import com.epmedu.animeal.feeding.presentation.ui.FeedingPhotoItem
+import com.epmedu.animeal.feeding.presentation.model.FeedingPhotoItem
 import com.epmedu.animeal.home.presentation.HomeScreenEvent
 import com.epmedu.animeal.home.presentation.model.CameraState
 import com.epmedu.animeal.home.presentation.viewmodel.HomeState
